Improvements to retrieval
Added progress indicator for downloads that are taking a long time.

Username and password aren’t sent unless they’re needed (the server
returning a 401).

If username and password are required (based on a 401 response) and
they weren’t supplied through the environment or command line
arguments, they’re prompted for interactively.